Move `Data` to `types.rs`
